refactor: 타입별 파일 분리
