+++ Calculator/Calculator/Model/CalculatorItemQueue.swift
@@ -43,86 +43,3 @@ struct CalculatorItemQueue<T: CalculateItem>: DoubleStackQueue {
     }
 }
 
-enum Operator: Character, CaseIterable, CalculateItem {
-    case add = "+"
-    case subtract = "-"
-    case divide = "/"
-    case multiply = "*"
-    
-    func calculate(lhs: Double, rhs: Double) -> Double {
-        switch self {
-        case .add:
-            return add(lhs: lhs, rhs: rhs)
-        case .subtract:
-            return subtract(lhs: lhs, rhs: rhs)
-        case .divide:
-            return divide(lhs: lhs, rhs: rhs)
-        case .multiply:
-            return multiply(lhs: lhs, rhs: rhs)
-        }
-    }
-    
-    private func add(lhs: Double, rhs: Double) -> Double {
-        return lhs + rhs
-    }
-    
-    private func subtract(lhs: Double, rhs: Double) -> Double {
-        return lhs - rhs
-    }
-    
-    private func divide(lhs: Double, rhs: Double) -> Double {
-        guard rhs.isZero == false else { return Double.nan }
-        return lhs / rhs
-    }
-    
-    private func multiply(lhs: Double, rhs: Double) -> Double {
-        return lhs * rhs
-    }
-}
-
-enum ExpressionParser {
-    static func parse(from input: String) -> Formula {
-        var formula = Formula()
-        let operands = componentsByOperators(from: input)
-        let operators = Array(input.filter { $0.isNumber == false })
-        
-        for operand in operands {
-            if let number = Double(operand) {
-                formula.operands.enqueue(number)
-            }
-        }
-        
-        for `operator` in operators {
-            if let oper = Operator(rawValue: `operator`) {
-                formula.operators.enqueue(oper)
-            }
-        }
-
-        return formula
-    }
-    
-    private static func componentsByOperators(from input: String) -> [String] {
-        return input.components(separatedBy: ["+", "-", "/", "*"])
-    }
-}
-
-struct Formula {
-    var operands: CalculatorItemQueue<Double> = CalculatorItemQueue()
-    var operators: CalculatorItemQueue<Operator> = CalculatorItemQueue()
-    
-    mutating func result() -> Double {
-        var result: Double = 0.0
-        guard var lhs: Double = operands.dequeue() else { return 0 }
-        
-        while operands.isEmpty == false {
-            let dequeueOperator = operators.dequeue()
-            
-            guard let rhs = operands.dequeue(),
-                  let calculatedValue = dequeueOperator?.calculate(lhs: lhs, rhs: rhs) else { return 0 }
-            
-            lhs = calculatedValue
-            result = calculatedValue
-        }
-        return result
-    }
-}
